feat: Latitude and longitude fixed

+++ Models/Suppliers.cs
@@ -10,8 +10,8 @@ namespace SheshaAPI.Models
         public string SupplierDescription { get; set; }
         public int SupplierRating { get; set; }
         public int SupplierReview { get; set; }
-        public float SupplierLatitude { get; set; }
-        public float SupplierLongitude { get; set; }
+        public decimal SupplierLatitude { get; set; }
+        public decimal SupplierLongitude { get; set; }
         public string SupplierImage { get; set; }
     }
 }
